+++ packages/telescope-posts/lib/client/templates/modules/post_admin.js
@@ -6,7 +6,7 @@ Template.post_admin.helpers({
     return !!Settings.get('requirePostsApproval') && this.status === Posts.config.STATUS_APPROVED;
   },
   shortScore: function(){
-    return Math.floor(this.score*1000)/1000;
+    return Math.floor(this.score*100)/100;
   }
 });
 
@@ -18,5 +18,21 @@ Template.post_admin.events({
   'click .unapprove-link': function(e){
     Meteor.call('unapprovePost', this._id);
     e.preventDefault();
+  },
+  'click .delete-link': function(e){
+    var post = this;
+
+    e.preventDefault();
+
+    if(confirm("Delete “"+post.title+"”?")){
+      Meteor.call("deletePostById", post._id, function(error) {
+        if (error) {
+          console.log(error);
+          Messages.flash(error.reason, 'error');
+        } else {
+          Messages.flash(i18n.t('your_post_has_been_deleted'), 'success');
+        }
+      });
+    }
   }
 });