+++ History.md
@@ -1,6 +1,8 @@
 ## v0.23.1
 
 * [BREAKING] Modules data context must now be passed on explicitely using the `moduleData` attribute. 
+* [BREAKING] Updated `category_title` template.
+* [BREAKING] Refactored `post_admin` template. Added delete link, now using icons, and moved post edit link from `post_info` to `post_admin`. 
 
 ## v0.23 “SubScope”